kafka_consumer/requirements.txt
kafka-python==1.4.7
redis==3.3.11
